% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{combined}
\alias{combined}
\title{AquaTK Toxicokinetic Data}
\format{
\subsection{\code{combined}}{

Combined daphnia and fish with 431 observations in total
}
}
\usage{
combined
}
\description{
Combined dataset
}
\keyword{datasets}
