% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{daphnia}
\alias{daphnia}
\title{Daphnia Toxicokinetic Data}
\format{
\subsection{\code{daphnia}}{

Daphnia TK data with 324 observations of internal / external concentration
}
}
\usage{
daphnia
}
\description{
Collected from literature by Collins
}
\keyword{datasets}
