% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fish}
\alias{fish}
\title{Fish Toxicokinetic Data}
\format{
\subsection{\code{fish}}{

Fish TK data with 107 observations of internal / external concentration
}
}
\usage{
fish
}
\description{
Collected from Stadnicka.
}
\keyword{datasets}
