% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joins.R
\name{join_lipid_fraction}
\alias{join_lipid_fraction}
\title{Join Lipid Fraction Estimates}
\usage{
join_lipid_fraction(combined)
}
\arguments{
\item{combined}{The AquaTK combined data frame}
}
\value{
A data frame.
}
\description{
Join Lipid Fraction Estimates
}
\examples{
join_lipid_fraction(AquaTK::combined)

}
