% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joins.R
\name{join_source_metadata}
\alias{join_source_metadata}
\title{Join Source Metadata}
\usage{
join_source_metadata(keys)
}
\arguments{
\item{keys}{A data frame of bibtex citekeys}
}
\value{
A data frame.
}
\description{
Join Source Metadata
}
\examples{
df <- data.frame(BIBTEXKEY = unique(AquaTK::combined$BIBTEXKEY))
join_source_metadata(df)

}
