% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sources}
\alias{sources}
\title{Source Metadata}
\format{
\subsection{\code{sources}}{

Data frame of source metadata, e.g., author, year, journal...
}
}
\usage{
sources
}
\description{
Additional data about the source of the measurements
}
\keyword{datasets}
